+++ .jshintrc
@@ -1,5 +1,4 @@
 {
-    "bitwise": true,
     "camelcase": true,
     "curly": true,
     "eqeqeq": true,
